Refactor weather functionality and add new services
Removed the WeatherForecastController and updated Program.cs to include new weather-related services. Introduced WeatherService and IWeatherClient interfaces for improved data retrieval, along with a FakeWeatherClient for testing. Added WeatherController to handle HTTP requests and WeatherRequest for request parameters. Implemented validation with WeatherRequestValidator and added OpenWeatherClient for real-time data integration. Updated WeatherForecastDto for precision and removed unused project references.

+++ Backend/Api/Application/Weather/DTO/WeatherForecastDto.cs
@@ -15,14 +15,14 @@ namespace Application.Weather.DTO
         public DateTime ToUtc { get; init; }
         public IReadOnlyList<WeatherSliceDto> Slices { get; init; } = Array.Empty<WeatherSliceDto>();
     }
-
+    
     public sealed class WeatherSliceDto
     {
         public DateTime TimeUtc { get; init; }
         public double TemperatureC { get; init; }
         public double WindSpeedMs { get; init; }
         public int HumidityPercent { get; init; }
-        public float rainVolumeMm { get; init; }
+        public double rainVolumeMm { get; init; }
         public string ConditionText { get; init; } = string.Empty;
         public string ConditionIconUrl { get; init; } = string.Empty;
         public string Source { get; init; } = "openweather";
